[Design] Login, SignUp page 작업 완료, Controller 들 Rename
