+++ src/main/java/com/pikapedia/game/GameC.java
@@ -1,4 +1,4 @@
-package com.pikapedia.account;
+package com.pikapedia.game;
 
 import java.io.IOException;
 import javax.servlet.ServletException;
@@ -7,21 +7,17 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-@WebServlet("/JoinAccountC")
-public class JoinAccountC extends HttpServlet {
+@WebServlet("/GameC")
+public class GameC extends HttpServlet {
+
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		AccountDAO.checkLogin(request);
-		request.setAttribute("contentPage", "jsp/signup.jsp");
+		
+		GameDAO.random(request);
+		request.setAttribute("contentPage", "jsp/quizOutput.jsp");
 		request.getRequestDispatcher("index.jsp").forward(request, response);
 	}
-
 	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		// 회원가입하는 일
-		AccountDAO.joinAccount(request);
-		// 어디로?
-		AccountDAO.checkLogin(request);
-		request.setAttribute("contentPage", "jsp/pokemonMain.jsp");
-		request.getRequestDispatcher("index.jsp").forward(request, response);
+	
 	}
 
 }
