+++ src/main/java/com/pikapedia/serach/SearchHC.java
@@ -11,7 +11,7 @@ import javax.servlet.http.HttpServletResponse;
 public class SearchHC extends HttpServlet {
        
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		request.getRequestDispatcher("input.jsp").forward(request, response);
+		request.getRequestDispatcher(".jsp").forward(request, response);
 	}
 
 	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
