+++ src/main/webapp/css/login.css
@@ -1,1 +1,56 @@
 @charset "UTF-8";
+
+#pokedex-white-bg{
+    display: flex;
+    justify-content: center;
+}
+.btn-container{
+    /*display: flex;*/
+    /*flex-direction: column;*/
+    align-items: center;
+    justify-content: center;
+    margin-top: 20vh;
+}
+.btn-container > div{
+    position: relative;
+}
+.info-input-btn{
+    width: 25vw;
+    height: 40px;
+    background-color: lightgray;
+    border-radius: 1.5vw;
+    border: transparent;
+    margin-bottom: 10px;
+    padding-left: 6vw;
+}
+.user-icon {
+    position: absolute;
+    width: 20px;
+    left: 2vw;
+    margin-top: 10px;
+}
+.confirm-btn-container{
+    display: flex;
+    justify-content: space-between;
+}
+.confirm-btn{
+    width: 12vw;
+    height: 40px;
+    line-height: 40px;
+    background-color: gray;
+    text-align: center;
+    border-radius: 1.5vw;
+    border: transparent;
+    color: white;
+}
+.google-logo-container{
+    display: flex;
+    justify-content: center;
+}
+.google-logo{
+    width: 25px;
+}
+.or-signup{
+    color: black;
+    text-align: center;
+}
